backend_fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
pydantic
email-validator
python-jose
passlib[bcrypt]
python-multipart
redis==5.0.1
redis>=4.2.0
pika
alembic